Move Source form component to a component-common dir.
